speed up initial import of default styles
Wrap the import of the supplied styles in a database transaction to
avoid flushing to disk after every single style is imported.  This
reduces the import time when the configdir is located on a HDD from
about three minutes to less than two seconds (and correspondingly
speeds up the first startup of a new installation).  Initial import
takes less than a second on an SSD even without the transaction,
so users with such a configuration won't notice any difference.